# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Recipe Management Service API
  description: |
    Comprehensive recipe management service for the Recipe Web Application ecosystem.

    Provides complete functionality for managing recipes, ingredients, steps, tags,
    reviews, and media content. Built with Spring Boot 3.5.3 and Java 21.

    **Features:**
    - Complete recipe CRUD operations
    - Advanced search and filtering
    - Ingredient management with scaling and shopping lists
    - Step-by-step cooking instructions
    - User reviews and ratings
    - Tag-based categorization
    - Media upload and management
    - Revision history tracking
    - JWT authentication integration
  version: 1.0.0
  contact:
    name: Recipe Web App Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1/recipe-management
    description: Local development server
  - url: http://sous-chef-proxy.local/api/v1/recipe-management
    description: Kubernetes deployment
security:
  - BearerAuth: []
...
paths:
  /recipes:
    get:
      tags: [recipes]
      summary: Get all recipes
      description: |
        Retrieves all recipes with pagination support. Returns basic recipe information
        including title, description, ratings, and metadata.
      operationId: getAllRecipes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: List of recipes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchRecipesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [recipes]
      summary: Create new recipe
      description: |
        Creates a new recipe with ingredients, steps, and tags. The recipe is
        associated with the authenticated user.
      operationId: createRecipe
      requestBody:
        description: Recipe creation data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
      responses:
        '201':
          description: Recipe created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /recipes/{recipeId}:
    get:
      tags: [recipes]
      summary: Get recipe by ID
      description: |
        Retrieves a complete recipe with all associated data including ingredients,
        steps, tags, and summary statistics.
      operationId: getRecipeById
      parameters:
        - $ref: '#/components/parameters/RecipeIdParam'
        - name: includeComments
          in: query
          required: false
          description: Include recipe comments in the response
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Recipe retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [recipes]
      summary: Update recipe
      description: |
        Updates an existing recipe. User must own the recipe or have appropriate
        permissions to modify it.
      operationId: updateRecipe
      parameters:
        - $ref: '#/components/parameters/RecipeIdParam'
      requestBody:
        description: Recipe update data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecipeRequest'
      responses:
        '200':
          description: Recipe updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [recipes]
      summary: Delete recipe
      description: |
        Permanently deletes a recipe and all associated data including ingredients,
        steps, reviews, and media. User must own the recipe to delete it.
      operationId: deleteRecipe
      parameters:
        - $ref: '#/components/parameters/RecipeIdParam'
      responses:
        '204':
          description: Recipe deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /favorites/recipes:
    get:
      tags: [favorites]
      summary: Get all favorited recipes
      description: |
        Retrieves all recipes favorited by a specified user with pagination support.
        Returns recipes in the same format as the main recipes list endpoint.

        **Privacy Controls:**
        This endpoint integrates with user-management-service to check user privacy
        preferences. If the target user's profile is private, only users they follow
        can retrieve their favorites list. Returns 403 Forbidden if the requesting
        user is not authorized to view the favorites.

        **Default Behavior:**
        If userId is not provided, returns favorites for the authenticated user.
      operationId: getFavoriteRecipes
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Favorited recipes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchRecipesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - User's favorites are private and requester is not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Forbidden"
                message: "User's favorites are private. Only followers can view them."
                timestamp: "2024-08-28T10:30:00Z"
                path: "/favorites/recipes"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /favorites/recipes/{recipeId}:
    post:
      tags: [favorites]
      summary: Favorite a recipe
      description: |
        Adds a recipe to the authenticated user's favorites.

        **Visibility:**
        Any authenticated user can favorite any recipe regardless of recipe ownership
        or visibility. There are no restrictions on which recipes can be favorited.

        Returns 409 Conflict if the recipe is already favorited by this user.
      operationId: favoriteRecipe
      parameters:
        - $ref: '#/components/parameters/RecipeIdParam'
      responses:
        '201':
          description: Recipe favorited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeFavoriteDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict - Recipe is already favorited by this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Conflict"
                message: "User has already favorited this recipe"
                timestamp: "2024-08-28T10:30:00Z"
                path: "/favorites/recipes/123"
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [favorites]
      summary: Unfavorite a recipe
      description: |
        Removes a recipe from the authenticated user's favorites.

        Returns 404 if the favorite does not exist (i.e., the recipe was not
        favorited by this user).
      operationId: unfavoriteRecipe
      parameters:
        - $ref: '#/components/parameters/RecipeIdParam'
      responses:
        '204':
          description: Recipe unfavorited successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Not Found - Recipe was not favorited by this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Not Found"
                message: "Favorite not found for this user and recipe"
                timestamp: "2024-08-28T10:30:00Z"
                path: "/favorites/recipes/123"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /favorites/recipes/{recipeId}/is-favorited:
    get:
      tags: [favorites]
      summary: Check if recipe is favorited
      description: |
        Checks whether the authenticated user has favorited the specified recipe.

        Returns a boolean value indicating favorite status.
      operationId: isRecipeFavorited
      parameters:
        - $ref: '#/components/parameters/RecipeIdParam'
      responses:
        '200':
          description: Successfully retrieved favorite status
          content:
            application/json:
              schema:
                type: boolean
              examples:
                favorited:
                  value: true
                  summary: Recipe is favorited
                notFavorited:
                  value: false
                  summary: Recipe is not favorited
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /favorites/collections:
    get:
      tags: [favorites]
      summary: Get all favorited collections
      description: |
        Retrieves all collections favorited by a specified user with pagination support.
        Returns collections in the same format as the main collections list endpoint.

        **Privacy Controls:**
        This endpoint integrates with user-management-service to check user privacy
        preferences. If the target user's profile is private, only users they follow
        can retrieve their favorites list. Returns 403 Forbidden if the requesting
        user is not authorized to view the favorites.

        **Default Behavior:**
        If userId is not provided, returns favorites for the authenticated user.
      operationId: getFavoriteCollections
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Favorited collections retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCollectionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - User's favorites are private and requester is not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Forbidden"
                message: "User's favorites are private. Only followers can view them."
                timestamp: "2024-08-28T10:30:00Z"
                path: "/favorites/collections"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /favorites/collections/{collectionId}:
    post:
      tags: [favorites]
      summary: Favorite a collection
      description: |
        Adds a collection to the authenticated user's favorites.

        **Visibility:**
        Users can only favorite collections they have access to view. This includes
        PUBLIC collections and collections where the user has been granted access
        (e.g., as a collaborator or through other permission mechanisms).

        Returns 403 Forbidden if the user does not have access to view the collection.
        Returns 409 Conflict if the collection is already favorited by this user.
      operationId: favoriteCollection
      parameters:
        - $ref: '#/components/parameters/CollectionIdParam'
      responses:
        '201':
          description: Collection favorited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionFavoriteDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - User does not have access to view this collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Forbidden"
                message: "You do not have access to view this collection"
                timestamp: "2024-08-28T10:30:00Z"
                path: "/favorites/collections/301"
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict - Collection is already favorited by this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Conflict"
                message: "User has already favorited this collection"
                timestamp: "2024-08-28T10:30:00Z"
                path: "/favorites/collections/301"
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [favorites]
      summary: Unfavorite a collection
      description: |
        Removes a collection from the authenticated user's favorites.

        Returns 404 if the favorite does not exist (i.e., the collection was not
        favorited by this user).
      operationId: unfavoriteCollection
      parameters:
        - $ref: '#/components/parameters/CollectionIdParam'
      responses:
        '204':
          description: Collection unfavorited successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Not Found - Collection was not favorited by this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Not Found"
                message: "Favorite not found for this user and collection"
                timestamp: "2024-08-28T10:30:00Z"
                path: "/favorites/collections/301"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /favorites/collections/{collectionId}/is-favorited:
    get:
      tags: [favorites]
      summary: Check if collection is favorited
      description: |
        Checks whether the authenticated user has favorited the specified collection.

        Returns a boolean value indicating favorite status.
      operationId: isCollectionFavorited
      parameters:
        - $ref: '#/components/parameters/CollectionIdParam'
      responses:
        '200':
          description: Successfully retrieved favorite status
          content:
            application/json:
              schema:
                type: boolean
              examples:
                favorited:
                  value: true
                  summary: Collection is favorited
                notFavorited:
                  value: false
                  summary: Collection is not favorited
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /collections:
    get:
      tags: [collections]
      summary: Get accessible collections
      description: |
        Retrieves all collections that the authenticated user can access based on
        permissions. This includes collections they own, collections they collaborate on,
        and public collections.
      operationId: getCollections
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Collections retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCollectionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [collections]
      summary: Create new collection
      description: |
        Creates a new recipe collection owned by the authenticated user.

        **Batch Operations:**
        - Optionally include `recipeIds` to add recipes to the collection during creation
        - Optionally include `collaboratorIds` to add collaborators during creation
          (only applicable when `collaborationMode` is `SPECIFIC_USERS`)

        This reduces multiple API calls to a single request when creating a collection
        with initial recipes and/or collaborators.
      operationId: createCollection
      requestBody:
        description: Collection creation data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionRequest'
      responses:
        '201':
          description: Collection created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /collections/{collectionId}:
    get:
      tags: [collections]
      summary: Get collection details
      description: |
        Retrieves complete collection details including all recipes in display order.
        User must have view permission for the collection.
      operationId: getCollectionById
      parameters:
        - $ref: '#/components/parameters/CollectionIdParam'
      responses:
        '200':
          description: Collection retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDetailsDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [collections]
      summary: Update collection
      description: |
        Updates collection metadata (name, description, visibility, collaboration mode).
        User must own the collection or have edit permission.
      operationId: updateCollection
      parameters:
        - $ref: '#/components/parameters/CollectionIdParam'
      requestBody:
        description: Collection update data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCollectionRequest'
      responses:
        '200':
          description: Collection updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [collections]
      summary: Delete collection
      description: |
        Permanently deletes a collection and all associated items and collaborators.
        User must own the collection to delete it.
      operationId: deleteCollection
      parameters:
        - $ref: '#/components/parameters/CollectionIdParam'
      responses:
        '204':
          description: Collection deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token authentication. Obtain token from user-management-service
        and include in Authorization header as 'Bearer <token>'.
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (0-based)
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      description: Page size
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: Sort criteria (property,direction)
      required: false
      schema:
        type: array
        items:
          type: string
        example: ["createdAt,desc", "title,asc"]
    RecipeIdParam:
      name: recipeId
      in: path
      description: Recipe identifier
      required: true
      schema:
        type: integer
        format: int64
        example: 123
...
  schemas:
...
    SearchRecipesResponse:
      type: object
      required:
        - content
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RecipeDto'
        totalElements:
          type: integer
          format: int64
          description: Total number of matching recipes
        totalPages:
          type: integer
          description: Total number of pages
        first:
          type: boolean
          description: Whether this is the first page
        last:
          type: boolean
          description: Whether this is the last page
        numberOfElements:
          type: integer
          description: Number of elements in current page
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Bad Request"
            message: "Invalid request parameters"
            timestamp: "2024-08-28T10:30:00Z"
            path: "/recipes"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "JWT token is missing or invalid"
            timestamp: "2024-08-28T10:30:00Z"
            path: "/recipes"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Not Found"
            message: "Recipe with ID 123 not found"
            timestamp: "2024-08-28T10:30:00Z"
            path: "/recipes/123"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Validation Error"
            message: "Request validation failed"
            timestamp: "2024-08-28T10:30:00Z"
            path: "/recipes"
            details:
              title: "Title is required"
              servings: "Servings must be positive"
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal Server Error"
            message: "An unexpected error occurred"
            timestamp: "2024-08-28T10:30:00Z"
            path: "/recipes"
...